editoast: add trait OSRDObject and Referable
