+++ editoast/src/models/errors/track_section_links.rs
@@ -79,7 +79,7 @@ mod tests {
         ));
         let errors = generate_errors(&infra_cache);
         assert_eq!(1, errors.len());
-        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "E".into());
+        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "E");
         let infra_error = InfraError::new_invalid_reference("link_error", "dst.track", obj_ref);
         assert_eq!(infra_error, errors[0]);
     }
@@ -94,10 +94,10 @@ mod tests {
         ));
         let errors = generate_errors(&infra_cache);
         assert_eq!(2, errors.len());
-        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "E".into());
+        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "E");
         let infra_error = InfraError::new_invalid_reference("link_error", "src.track", obj_ref);
         assert_eq!(infra_error, errors[0]);
-        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "F".into());
+        let obj_ref = ObjectRef::new(ObjectType::TrackSection, "F");
         let infra_error = InfraError::new_invalid_reference("link_error", "dst.track", obj_ref);
         assert_eq!(infra_error, errors[1]);
     }
