+++ editoast/src/railjson/speed_section.rs
@@ -1,0 +1,29 @@
+use std::collections::HashMap;
+
+use super::generate_id;
+use super::ApplicableDirectionsTrackRange;
+use super::OSRDObject;
+use super::ObjectType;
+use derivative::Derivative;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
+#[derivative(Default)]
+pub struct SpeedSection {
+    #[derivative(Default(value = r#"generate_id("speed_section")"#))]
+    pub id: String,
+    #[derivative(Default(value = "Some(80.)"))]
+    pub speed_limit: Option<f64>,
+    pub speed_limit_by_tag: HashMap<String, f64>,
+    pub track_ranges: Vec<ApplicableDirectionsTrackRange>,
+}
+
+impl OSRDObject for SpeedSection {
+    fn get_id(&self) -> String {
+        self.id.clone()
+    }
+    fn get_type(&self) -> ObjectType {
+        ObjectType::SpeedSection
+    }
+}
