+++ editoast/src/railjson/switch_type.rs
@@ -1,0 +1,35 @@
+use std::collections::HashMap;
+
+use super::generate_id;
+use super::OSRDObject;
+use super::ObjectType;
+use derivative::Derivative;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize)]
+#[serde(deny_unknown_fields)]
+#[derivative(Default)]
+pub struct SwitchType {
+    #[derivative(Default(value = r#"generate_id("switchtype")"#))]
+    pub id: String,
+    pub ports: Vec<String>,
+    pub groups: HashMap<String, Vec<SwitchPortConnection>>,
+}
+
+impl OSRDObject for SwitchType {
+    fn get_id(&self) -> String {
+        self.id.clone()
+    }
+    fn get_type(&self) -> ObjectType {
+        ObjectType::SwitchType
+    }
+}
+
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Eq, Hash)]
+#[serde(deny_unknown_fields)]
+#[derivative(Default)]
+pub struct SwitchPortConnection {
+    pub src: String,
+    pub dst: String,
+    pub bidirectional: bool,
+}
